Refactor: 수익지출 관리 페이지 UI 수정
